requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.9.0
pandas>=1.3.0
webdriver-manager==4.0.1
python-multipart==0.0.6
gunicorn==20.1.0
numpy==1.23.5
undetected-chromedriver==3.5.5
chromedriver-py>=114.0.5735.90
python-dotenv==1.0.0
playwright==1.40.0
nest_asyncio==1.5.1
exif==1.6.0
asyncio==3.4.3
jinja2==3.1.2
aiofiles==23.2.1
twikit>=0.1.0
